+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release Notes
 
+## WIP
+
+- Batch update collection items state
+
 ## 2.3.6 (2023-01-03)
 
 - Fix tree model nesting items